+++ MISO/Modules/MISO/AppServices/Sources/EasterEggs/Confettis/Internal/ConfettiConfiguration.swift
@@ -1,0 +1,10 @@
+// Software: MISO iOS
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) Pierre-Yves Lapersonne
+// ✿✿✿✿ ʕ •ᴥ•ʔ/ ︻デ═一
+
+enum ConfettiConfiguration {
+
+    static let ANIMATION_DURATION: Double = 3.0
+    static let PARTICLE_COUNT = 120
+}